extract converting methods to their origins
